# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Miracle Birds API
  description: |
    # Miracle Birds - AI Intelligence Layer for CRM
    
    The Secure AI Intelligence Layer for Every CRM. Miracle Birds provides AI-powered customer intelligence, predictive analytics, and workflow automation on top of existing CRM platforms.
    
    ## Features
    - 🤖 AI CRM Copilot
    - 📊 Customer 360 Intelligence
    - 📈 Predictive Analytics (Churn, Lead Scoring, Revenue)
    - 🔗 CRM Integration (Salesforce, Zoho, HubSpot, Dynamics, Pipedrive)
    - 🔒 Enterprise Security & Compliance
    - 🚀 Workflow Automation
    
    ## Authentication
    All API endpoints (except auth endpoints) require a valid JWT token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Rate Limiting
    - **Standard Plan**: 1,000 requests/hour
    - **Professional Plan**: 5,000 requests/hour
    - **Enterprise Plan**: Unlimited
    
    ## Base URLs
    - **Production**: `https://api.miraclebirds.ai`
    - **Staging**: `https://api-staging.miraclebirds.ai`
    - **Development**: `http://localhost:8000`
    
  version: 1.0.0
  contact:
    name: Miracle Birds API Support
    email: api@miraclebirds.ai
    url: https://miraclebirds.ai/support
  license:
    name: Proprietary
    url: https://miraclebirds.ai/terms
servers:
  - url: https://api.miraclebirds.ai
    description: Production server
  - url: https://api-staging.miraclebirds.ai
    description: Staging server
...
paths:
...
  /api/v1/auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create a new user account
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - first_name
                - last_name
              properties:
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: SecurePass123!
                first_name:
                  type: string
                  example: John
                last_name:
                  type: string
                  example: Doe
                company_name:
                  type: string
                  example: Acme Corp
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return JWT tokens
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  format: password
                  example: SecurePass123!
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get new access token using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Invalidate user tokens
      operationId: logoutUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logged out successfully

  # ==================== CUSTOMERS ====================
...
  /api/v1/copilot/chat:
    post:
      tags:
        - AI Copilot
      summary: Chat with AI Copilot
      description: Send message to AI assistant
      operationId: chatWithCopilot
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  example: What are my top at-risk customers?
                context:
                  type: object
                  description: Additional context
      responses:
        '200':
          description: AI response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CopilotResponse'
  /api/v1/copilot/conversations:
    get:
      tags:
        - AI Copilot
      summary: List conversations
      description: Get chat history
      operationId: listConversations
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of conversations
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Conversation'

  # ==================== USERS ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid request
        detail:
          type: string
          example: Email is required
        code:
          type: string
          example: validation_error
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          example: 1800
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        role:
          type: string
          enum: [super_admin, admin, manager, user, viewer]
        tenant_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
...
    CopilotResponse:
      type: object
      properties:
        message:
          type: string
        conversation_id:
          type: string
          format: uuid
        suggestions:
          type: array
          items:
            type: string
    Conversation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        created_at:
          type: string
          format: date-time
        messages_count:
          type: integer
...